use bigserials for primary keys, add formal support for bigserial id types
Doing this required us to add an explicit IdType alias to use in our generators and throughout our app, which would point to the values kysely is expected for generated bigint types.

+++ test-app/db/migrations/000000002-create-compositions.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('compositions')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('user_id', 'integer', col => col.references('users.id').onDelete('cascade').notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('user_id', 'bigint', col => col.references('users.id').onDelete('cascade').notNull())
     .addColumn('content', 'text')
     .addColumn('primary', 'boolean', col => col.defaultTo(false))
     .addColumn('created_at', 'timestamp', col => col.notNull())
