+++ test-app/db/migrations/000000005-create-posts.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('posts')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('user_id', 'integer', col => col.references('users.id').onDelete('cascade').notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('user_id', 'bigint', col => col.references('users.id').onDelete('cascade').notNull())
     .addColumn('post_visibility_id', 'bigint', col => col.references('posts.id').onDelete('cascade'))
     .addColumn('body', 'text')
     .addColumn('created_at', 'timestamp', col => col.notNull())
