+++ test-app/db/migrations/000000007-create-user-settings.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('user_settings')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('user_id', 'integer', col => col.references('users.id').onDelete('cascade').notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('user_id', 'bigint', col => col.references('users.id').onDelete('cascade').notNull())
     .addColumn('likes_chalupas', 'boolean', col => col.notNull().defaultTo(false))
     .addColumn('created_at', 'timestamp', col => col.notNull())
     .addColumn('updated_at', 'timestamp', col => col.notNull())
