+++ test-app/db/migrations/1683311290078-create-incompatible-foreign-key-types.ts
@@ -3,8 +3,8 @@ import { Kysely, sql } from 'kysely'
 export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('incompatible_foreign_key_type_examples')
-    .addColumn('id', 'serial', col => col.primaryKey())
-    .addColumn('user_id', 'bigint', col => col.references('users.id').onDelete('cascade').notNull())
+    .addColumn('id', 'bigserial', col => col.primaryKey())
+    .addColumn('user_id', 'integer', col => col.references('users.id').onDelete('cascade').notNull())
     .addColumn('created_at', 'timestamp', col => col.notNull())
     .addColumn('updated_at', 'timestamp', col => col.notNull())
     .execute()
